wr octree

+++ templates/wr/src/Core/Renderer/Framework/cVertexLayout.h
@@ -38,7 +38,7 @@ template<>
 inline void cVertexLayout::push<float>( unsigned int _count )
 {
 	size_t size = sizeof( float );
-    m_stride += size * _count;
+    m_stride += (unsigned int)size * _count;
 	m_elements.push_back( { eType::Float, _count, size } );
 }
 
@@ -46,7 +46,7 @@ template<>
 inline void cVertexLayout::push<unsigned char>( unsigned int _count )
 {
 	size_t size = sizeof( unsigned char );
-	m_stride += size * _count;
+	m_stride += (unsigned int)size * _count;
 	m_elements.push_back( { eType::Byte, _count, size } );
 }
 
@@ -54,7 +54,7 @@ template<>
 inline void cVertexLayout::push<unsigned int>( unsigned int _count )
 {
 	size_t size = sizeof( unsigned int );
-	m_stride += size * _count;
+	m_stride += (unsigned int)size * _count;
 	m_elements.push_back( { eType::Uint, _count, size } );
 }
 
@@ -62,6 +62,6 @@ template<>
 inline void cVertexLayout::push<double>( unsigned int _count )
 {
 	size_t size = sizeof( double );
-	m_stride += size * _count;
+	m_stride += (unsigned int)size * _count;
 	m_elements.push_back( { eType::Double, _count, size } );
 }
